test, refactor: 기본 테스트 test container test로 수정
